App-3/requirements.txt
Flask==2.3.2
google.auth==2.22.0
requests==2.31.0